+++ src/specify_cli/commands/__init__.py
@@ -1,0 +1,1 @@
+"""CLI command groups — each module exposes a register(app) function."""
