+++ src/specify_cli/commands/extension.py
@@ -1,0 +1,1 @@
+"""specify extension * commands — placeholder for future extraction."""
