+++ src/specify_cli/commands/integration.py
@@ -1,0 +1,1 @@
+"""specify integration * commands — placeholder for future extraction."""
