+++ src/specify_cli/commands/preset.py
@@ -1,0 +1,1 @@
+"""specify preset * commands — placeholder for future extraction."""
