+++ src/specify_cli/commands/workflow.py
@@ -1,0 +1,1 @@
+"""specify workflow * commands — placeholder for future extraction."""
